backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
python-decouple==3.8
django-allauth==65.11.2
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
